+++ backend/pkg/jwt/custom_claims.go
@@ -1,4 +1,4 @@
-package utils
+package jwt
 
 import "github.com/golang-jwt/jwt/v5"
