+++ backend/pkg/jwt/jwt_generator.go
@@ -1,4 +1,4 @@
-package utils
+package jwt
 
 import (
 	"fmt"
