+++ backend/pkg/jwt/jwt_verifier.go
@@ -1,4 +1,4 @@
-package utils
+package jwt
 
 import (
 	"errors"
